+++ pennylane_lightning/lightning_gpu/_state_vector.py
@@ -229,34 +229,39 @@ class LightningGPUStateVector(LightningBaseStateVector):
         # set the state vector on GPU with provided state and their corresponding wires
         self._qubit_state.setStateVector(state, list(device_wires), use_async)
 
-    def _apply_lightning_controlled(self, operation):
+    def _apply_lightning_controlled(self, operation, adjoint):
         """Apply an arbitrary controlled operation to the state tensor.
 
         Args:
             operation (~pennylane.operation.Operation): controlled operation to apply
+            adjoint (bool): Apply the adjoint of the operation if True
 
         Returns:
             None
         """
         state = self.state_vector
 
-        basename = operation.base.name
-        method = getattr(state, f"{basename}", None)
+        if isinstance(operation.base, Adjoint):
+            base_operation = operation.base.base
+            adjoint = not adjoint
+        else:
+            base_operation = operation.base
+
+        method = getattr(state, f"{base_operation.name}", None)
         control_wires = list(operation.control_wires)
         control_values = operation.control_values
         target_wires = list(operation.target_wires)
-        if method:  # apply n-controlled specialized gate
-            inv = False
+        if method is not None:  # apply n-controlled specialized gate
             param = operation.parameters
-            method(control_wires, control_values, target_wires, inv, param)
+            method(control_wires, control_values, target_wires, adjoint, param)
         else:  # apply gate as an n-controlled matrix
             method = getattr(state, "applyControlledMatrix")
             method(
-                qml.matrix(operation.base),
+                qml.matrix(base_operation),
                 control_wires,
                 control_values,
                 target_wires,
-                False,
+                adjoint,
             )
 
     def _apply_lightning_midmeasure(
@@ -289,7 +294,10 @@ class LightningGPUStateVector(LightningBaseStateVector):
 
     # pylint: disable=unused-argument, too-many-branches
     def _apply_lightning(
-        self, operations, mid_measurements: dict = None, postselect_mode: str = None
+        self,
+        operations,
+        mid_measurements: dict = None,
+        postselect_mode: str = None,
     ):
         """Apply a list of operations to the state vector.
 
@@ -299,6 +307,7 @@ class LightningGPUStateVector(LightningBaseStateVector):
             postselect_mode (str): Configuration for handling shots with mid-circuit measurement
                 postselection. Use ``"hw-like"`` to discard invalid shots and ``"fill-shots"`` to
                 keep the same number of shots. Default is ``None``.
+
 
         Returns:
             None
@@ -330,9 +339,12 @@ class LightningGPUStateVector(LightningBaseStateVector):
                 param = operation.parameters
                 method(wires, invert_param, param)
             elif (
-                isinstance(operation, qml.ops.Controlled) and not self._mpi_handler.use_mpi
+                isinstance(operation.base if invert_param else operation, qml.ops.Controlled)
+                and not self._mpi_handler.use_mpi
             ):  # MPI backend does not have native controlled gates support
-                self._apply_lightning_controlled(operation)
+                self._apply_lightning_controlled(
+                    operation.base if invert_param else operation, invert_param
+                )
             elif (
                 self._mpi_handler.use_mpi
                 and isinstance(operation, qml.ops.Controlled)
@@ -348,7 +360,6 @@ class LightningGPUStateVector(LightningBaseStateVector):
                 except AttributeError:  # pragma: no cover
                     # To support older versions of PL
                     mat = operation.matrix
-
                 r_dtype = np.float32 if self.dtype == np.complex64 else np.float64
                 param = (
                     [[r_dtype(operation.hash)]]